Refactored context into its own module
