+++ bumpversion/utils.py
@@ -1,15 +1,7 @@
 """General utilities."""
 
-import datetime
 import string
-from collections import ChainMap
-from dataclasses import asdict
-from typing import TYPE_CHECKING, Any, List, Optional, Tuple
-
-if TYPE_CHECKING:  # pragma: no-coverage
-    from bumpversion.config import Config
-    from bumpversion.scm import SCMInfo
-    from bumpversion.versioning.models import Version
+from typing import Any, List, Tuple
 
 
 def extract_regex_flags(regex_pattern: str) -> Tuple[str, str]:
@@ -42,46 +34,9 @@ def key_val_string(d: dict) -> str:
     return ", ".join(f"{k}={v}" for k, v in sorted(d.items()))
 
 
-def prefixed_environ() -> dict:
-    """Return a dict of the environment with keys wrapped in `${}`."""
-    import os
-
-    return {f"${key}": value for key, value in os.environ.items()}
-
-
 def labels_for_format(serialize_format: str) -> List[str]:
     """Return a list of labels for the given serialize_format."""
     return [item[1] for item in string.Formatter().parse(serialize_format) if item[1]]
-
-
-def base_context(scm_info: Optional["SCMInfo"] = None) -> ChainMap:
-    """The default context for rendering messages and tags."""
-    from bumpversion.scm import SCMInfo  # Including this here to avoid circular imports
-
-    scm = asdict(scm_info) if scm_info else asdict(SCMInfo())
-
-    return ChainMap(
-        {
-            "now": datetime.datetime.now(),
-            "utcnow": datetime.datetime.now(datetime.timezone.utc),
-        },
-        prefixed_environ(),
-        scm,
-        {c: c for c in ("#", ";")},
-    )
-
-
-def get_context(
-    config: "Config", current_version: Optional["Version"] = None, new_version: Optional["Version"] = None
-) -> ChainMap:
-    """Return the context for rendering messages and tags."""
-    ctx = base_context(config.scm_info)
-    ctx = ctx.new_child({"current_version": config.current_version})
-    if current_version:
-        ctx = ctx.new_child({f"current_{part}": current_version[part].value for part in current_version})
-    if new_version:
-        ctx = ctx.new_child({f"new_{part}": new_version[part].value for part in new_version})
-    return ctx
 
 
 def get_overrides(**kwargs) -> dict:
