+++ bumpversion/versioning/functions.py
@@ -4,31 +4,7 @@ import datetime
 import re
 from typing import List, Optional, Union
 
-
-def get_datetime_info(current_dt: datetime.datetime) -> dict:
-    """Return the full structure of the given datetime for formatting."""
-    return {
-        "YYYY": current_dt.strftime("%Y"),
-        "YY": current_dt.strftime("%y").lstrip("0") or "0",
-        "0Y": current_dt.strftime("%y"),
-        "MMM": current_dt.strftime("%b"),
-        "MM": str(current_dt.month),
-        "0M": current_dt.strftime("%m"),
-        "DD": str(current_dt.day),
-        "0D": current_dt.strftime("%d"),
-        "JJJ": current_dt.strftime("%j").lstrip("0"),
-        "00J": current_dt.strftime("%j"),
-        "Q": str((current_dt.month - 1) // 3 + 1),
-        "WW": current_dt.strftime("%W").lstrip("0") or "0",
-        "0W": current_dt.strftime("%W"),
-        "UU": current_dt.strftime("%U").lstrip("0") or "0",
-        "0U": current_dt.strftime("%U"),
-        "VV": current_dt.strftime("%V").lstrip("0") or "0",
-        "0V": current_dt.strftime("%V"),
-        "GGGG": current_dt.strftime("%G"),
-        "GG": current_dt.strftime("%G")[2:].lstrip("0") or "0",
-        "0G": current_dt.strftime("%G")[2:],
-    }
+from bumpversion.context import get_datetime_info
 
 
 class PartFunction:
